Changed the data type of ROM/RAM banks

+++ gb++/src/Mapper/MBC0/MBC0.cpp
@@ -1,6 +1,6 @@
 #include "MBC0.hpp"
 
-MBC0::MBC0(uint8_t NumROMBanks, uint8_t NumRAMBanks) : Mapper(NumROMBanks, NumRAMBanks) {
+MBC0::MBC0(uint16_t NumROMBanks, uint16_t NumRAMBanks) : Mapper(NumROMBanks, NumRAMBanks) {
 }
 
 MBC0::~MBC0() {
