+++ gb++/src/Mapper/Mapper.cpp
@@ -1,6 +1,6 @@
 #include "Mapper.hpp"
 
-Mapper::Mapper(uint8_t NumROMBanks, uint8_t NumRAMBanks) {
+Mapper::Mapper(uint16_t NumROMBanks, uint16_t NumRAMBanks) {
 	ROMBanks = NumROMBanks;
 	RAMBanks = NumRAMBanks;
 }
